Fix regex to remove DeprecationWarning
